Update main.go
Added system to calculate transaction reception latency by the difference between the timestamp sent by Shredstream Decoder (when the transaction is deserialized) and the listener server timestamp when it is received. This implementation is useful to find out how long it takes for the transaction from Shredstream Decoder to reach our listener when is hosted on another machine/datacenter.